requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMuPDF==1.23.8
Werkzeug==3.0.1
google-genai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
supabase>=2.0.0
tenacity>=8.2.0
PyPDF2>=3.0.0
Pillow>=10.0.0
easyocr>=1.6.0
google-cloud-vision>=3.0.0
opencv-python>=4.8.0
pdf2image>=1.16.0
